update Gallery in readme
This updates the `## Gallery` section in the readme file.
It adds screenshots of how the rosepine, rosepine-moon and rosepine-dawn
themes would look like in a codeblock.

As I don't know how I would generate the same image as was previously in
the readme file for the other themes I decided to replace all with new
screenshots.